レビュー修正(エラーハンドリング)

+++ app/controllers/products_controller.rb
@@ -24,8 +24,12 @@ class ProductsController < ApplicationController
 
   def destroy
     product = Product.find(params[:id])
-    product.destroy
-    redirect_to root_path
+    if current_user.id == product.user.id
+      product.destroy
+      redirect_to root_path
+    else
+      redirect_to root_path
+    end
   end
 
   private
